feature: add SearchShakeUseCase - use specification  in jpa repository - use performance data structure to manage creation of shake view

+++ src/main/java/com/lps/vitalMagic/shake/application/view/ShakeAttributeView.java
@@ -3,6 +3,6 @@ package com.lps.vitalMagic.shake.application.view;
 public record ShakeAttributeView(
         Long attributeId,
         String attributeName,
-        Byte value
+        int value
 ) {
 }
